2025/retry/pyproject.toml
[project]
name = "retry"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
]
